extract ResourceClusterListAdvancedRefreshFunc
