+++ internal/service/advancedclustertpf/common_test.go
@@ -1,13 +1,134 @@
 package advancedclustertpf_test
 
 import (
+	"errors"
+	"net/http"
 	"testing"
 
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/advancedclustertpf"
 	"github.com/stretchr/testify/assert"
-
+	"github.com/stretchr/testify/mock"
 	"go.mongodb.org/atlas-sdk/v20250312007/admin"
+	"go.mongodb.org/atlas-sdk/v20250312007/mockadmin"
 )
+
+var (
+	dummyProjectID   = "projectId"
+	errGeneric       = errors.New("generic")
+	advancedClusters = []admin.ClusterDescription20240805{{StateName: conversion.StringPtr("NOT IDLE")}}
+)
+
+type Result struct {
+	response any
+	error    error
+	state    string
+}
+
+func TestResourceListAdvancedRefreshFunc(t *testing.T) {
+	testCases := []struct {
+		mockCluster    *admin.PaginatedClusterDescription20240805
+		mockResponse   *http.Response
+		expectedResult Result
+		mockError      error
+		name           string
+		expectedError  bool
+	}{
+		{
+			name:          "Error in the API call: reset by peer",
+			mockError:     errors.New("reset by peer"),
+			expectedError: false,
+			expectedResult: Result{
+				response: nil,
+				state:    "REPEATING",
+				error:    nil,
+			},
+		},
+		{
+			name:          "Generic error in the API call",
+			mockError:     errGeneric,
+			expectedError: true,
+			expectedResult: Result{
+				response: nil,
+				state:    "",
+				error:    errGeneric,
+			},
+		},
+		{
+			name:          "Error in the API call: HTTP 404",
+			mockError:     errGeneric,
+			mockResponse:  &http.Response{StatusCode: http.StatusNotFound},
+			expectedError: false,
+			expectedResult: Result{
+				response: "",
+				state:    "DELETED",
+				error:    nil,
+			},
+		},
+		{
+			name:          "Error in the API call: HTTP 503",
+			mockError:     errGeneric,
+			mockResponse:  &http.Response{StatusCode: http.StatusServiceUnavailable},
+			expectedError: false,
+			expectedResult: Result{
+				response: "",
+				state:    "PENDING",
+				error:    nil,
+			},
+		},
+		{
+			name:          "Error in the API call: Neither HTTP 503 or 404",
+			mockError:     errGeneric,
+			mockResponse:  &http.Response{StatusCode: http.StatusBadRequest},
+			expectedError: true,
+			expectedResult: Result{
+				response: nil,
+				state:    "",
+				error:    errGeneric,
+			},
+		},
+		{
+			name:          "Successful but with at least one cluster not idle",
+			mockCluster:   &admin.PaginatedClusterDescription20240805{Results: &advancedClusters},
+			mockResponse:  &http.Response{StatusCode: http.StatusOK},
+			expectedError: false,
+			expectedResult: Result{
+				response: advancedClusters[0],
+				state:    "PENDING",
+				error:    nil,
+			},
+		},
+		{
+			name:          "Successful",
+			mockCluster:   &admin.PaginatedClusterDescription20240805{},
+			mockResponse:  &http.Response{StatusCode: http.StatusOK},
+			expectedError: false,
+			expectedResult: Result{
+				response: &admin.PaginatedClusterDescription20240805{},
+				state:    "IDLE",
+				error:    nil,
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			testObject := mockadmin.NewClustersApi(t)
+
+			testObject.EXPECT().ListClusters(mock.Anything, mock.Anything).Return(admin.ListClustersApiRequest{ApiService: testObject}).Once()
+			testObject.EXPECT().ListClustersExecute(mock.Anything).Return(tc.mockCluster, tc.mockResponse, tc.mockError).Once()
+
+			result, stateName, err := advancedclustertpf.ResourceClusterListAdvancedRefreshFunc(t.Context(), dummyProjectID, testObject)()
+			if (err != nil) != tc.expectedError {
+				t.Errorf("Case %s: Received unexpected error: %v", tc.name, err)
+			}
+
+			assert.Equal(t, tc.expectedResult.error, err)
+			assert.Equal(t, tc.expectedResult.response, result)
+			assert.Equal(t, tc.expectedResult.state, stateName)
+		})
+	}
+}
 
 func TestAddIDsToReplicationSpecs(t *testing.T) {
 	testCases := map[string]struct {
