+++ internal/service/privateendpointregionalmode/resource_private_endpoint_regional_mode.go
@@ -11,7 +11,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/validate"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/advancedcluster"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/advancedclustertpf"
 	"go.mongodb.org/atlas-sdk/v20250312007/admin"
 )
 
@@ -112,11 +112,9 @@ func resourceUpdate(ctx context.Context, d *schema.ResourceData, meta any) diag.
 	log.Println("[INFO] Waiting for MongoDB Clusters' Private Endpoints to be updated")
 
 	stateConf := &retry.StateChangeConf{
-		Pending: []string{"REPEATING", "PENDING"},
-		Target:  []string{"IDLE", "DELETED"},
-		// TODO: TEMPORARY CHANGE, DON'T MERGE
-		// TODO: TEMPORARY CHANGE, DON'T MERGE
-		Refresh:    advancedcluster.ResourceClusterListAdvancedRefreshFunc(ctx, projectID, conn.ClustersApi),
+		Pending:    []string{"REPEATING", "PENDING"},
+		Target:     []string{"IDLE", "DELETED"},
+		Refresh:    advancedclustertpf.ResourceClusterListAdvancedRefreshFunc(ctx, projectID, conn.ClustersApi),
 		Timeout:    d.Timeout(timeoutKey.(string)),
 		MinTimeout: 15 * time.Second,
 		Delay:      30 * time.Second, // give time for cluster connection strings to be updated
